+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Flows tab: fixed redirection for derivative dataset when list flows is empty
+
+
 ## [0.41.1] - 2025-03-07
 ### Fixed
 - `/page-not-found` route hierarchy